tests/wpt/tests/css/css-flexbox/animation/flex-shrink-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>flex-shrink interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#flex-shrink-property">
<meta name="assert" content="flex-shrink supports animation as a number">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.parent {
  flex-shrink: 3;
}
.target {
  flex-shrink: 1.5;
}
</style>

<body></body>

<script>
test_interpolation({
  property: 'flex-shrink',
  from: neutralKeyframe,
  to: '2',
}, [
  {at: -0.3, expect: '1.35'},
  {at: 0, expect: '1.5'},
  {at: 0.3, expect: '1.65'},
  {at: 0.6, expect: '1.8'},
  {at: 1, expect: '2'},
  {at: 1.5, expect: '2.25'},
]);

test_interpolation({
  property: 'flex-shrink',
  from: 'initial',
  to: '2',
}, [
  {at: -0.3, expect: '0.7'},
  {at: 0, expect: '1'},
  {at: 0.3, expect: '1.3'},
  {at: 0.6, expect: '1.6'},
  {at: 1, expect: '2'},
  {at: 1.5, expect: '2.5'},
]);

test_interpolation({
  property: 'flex-shrink',
  from: 'inherit',
  to: '2',
}, [
  {at: -0.3, expect: '3.3'},
  {at: 0, expect: '3'},
  {at: 0.3, expect: '2.7'},
  {at: 0.6, expect: '2.4'},
  {at: 1, expect: '2'},
  {at: 1.5, expect: '1.5'},
]);

test_interpolation({
  property: 'flex-shrink',
  from: 'unset',
  to: '2',
}, [
  {at: -0.3, expect: '0.7'},
  {at: 0, expect: '1'},
  {at: 0.3, expect: '1.3'},
  {at: 0.6, expect: '1.6'},
  {at: 1, expect: '2'},
  {at: 1.5, expect: '2.5'},
]);

test_interpolation({
  property: 'flex-shrink',
  from: '1',
  to: '2',
}, [
  {at: -5, expect: '0'},
  {at: -0.3, expect: '0.7'},
  {at: 0, expect: '1'},
  {at: 0.3, expect: '1.3'},
  {at: 0.6, expect: '1.6'},
  {at: 1, expect: '2'},
  {at: 1.5, expect: '2.5'},
]);

test_interpolation({
  property: 'flex-shrink',
  from: '0',
  to: '1',
}, [
  {at: -5, expect: '0'},
  {at: -0.3, expect: '0'},
  {at: 0, expect: '0'},
  {at: 0.3, expect: '0.3'},
  {at: 0.6, expect: '0.6'},
  {at: 1, expect: '1'},
  {at: 1.5, expect: '1.5'},
]);
</script>
