LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/animation/flex-shrink-interpolation-expected.txt


FAIL CSS Transitions: property <flex-shrink> from neutral to [2] at (-0.3) should be [1.35] assert_equals: expected "1.35 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from neutral to [2] at (0) should be [1.5] assert_equals: expected "1.5 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from neutral to [2] at (0.3) should be [1.65] assert_equals: expected "1.65 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from neutral to [2] at (0.6) should be [1.8] assert_equals: expected "1.8 " but got "2 "
PASS CSS Transitions: property <flex-shrink> from neutral to [2] at (1) should be [2]
FAIL CSS Transitions: property <flex-shrink> from neutral to [2] at (1.5) should be [2.25] assert_equals: expected "2.25 " but got "2 "
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (-0.3) should be [1.35]
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (0) should be [1.5]
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (0.3) should be [1.65]
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (0.6) should be [1.8]
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (1) should be [2]
PASS CSS Transitions with transition: all: property <flex-shrink> from neutral to [2] at (1.5) should be [2.25]
FAIL CSS Animations: property <flex-shrink> from neutral to [2] at (-0.3) should be [1.35] assert_equals: expected "1.35 " but got "1.5 "
PASS CSS Animations: property <flex-shrink> from neutral to [2] at (0) should be [1.5]
FAIL CSS Animations: property <flex-shrink> from neutral to [2] at (0.3) should be [1.65] assert_equals: expected "1.65 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from neutral to [2] at (0.6) should be [1.8] assert_equals: expected "1.8 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from neutral to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from neutral to [2] at (1.5) should be [2.25] assert_equals: expected "2.25 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from neutral to [2] at (-0.3) should be [1.35] assert_equals: expected "1.35 " but got "1.5 "
PASS Web Animations: property <flex-shrink> from neutral to [2] at (0) should be [1.5]
FAIL Web Animations: property <flex-shrink> from neutral to [2] at (0.3) should be [1.65] assert_equals: expected "1.65 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from neutral to [2] at (0.6) should be [1.8] assert_equals: expected "1.8 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from neutral to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from neutral to [2] at (1.5) should be [2.25] assert_equals: expected "2.25 " but got "1.5 "
FAIL CSS Transitions: property <flex-shrink> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [initial] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <flex-shrink> from [initial] to [2] at (1) should be [2]
FAIL CSS Transitions: property <flex-shrink> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (-0.3) should be [0.7]
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (0) should be [1]
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (0.3) should be [1.3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (0.6) should be [1.6]
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (1) should be [2]
PASS CSS Transitions with transition: all: property <flex-shrink> from [initial] to [2] at (1.5) should be [2.5]
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL CSS Transitions: property <flex-shrink> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "2 "
PASS CSS Transitions: property <flex-shrink> from [inherit] to [2] at (1) should be [2]
FAIL CSS Transitions: property <flex-shrink> from [inherit] to [2] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "2 "
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (-0.3) should be [3.3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (0) should be [3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (0.3) should be [2.7]
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (0.6) should be [2.4]
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (1) should be [2]
PASS CSS Transitions with transition: all: property <flex-shrink> from [inherit] to [2] at (1.5) should be [1.5]
FAIL CSS Animations: property <flex-shrink> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [inherit] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
PASS CSS Animations: property <flex-shrink> from [inherit] to [2] at (1.5) should be [1.5]
FAIL Web Animations: property <flex-shrink> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [inherit] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
PASS Web Animations: property <flex-shrink> from [inherit] to [2] at (1.5) should be [1.5]
FAIL CSS Transitions: property <flex-shrink> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [unset] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <flex-shrink> from [unset] to [2] at (1) should be [2]
FAIL CSS Transitions: property <flex-shrink> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (-0.3) should be [0.7]
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (0) should be [1]
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (0.3) should be [1.3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (0.6) should be [1.6]
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (1) should be [2]
PASS CSS Transitions with transition: all: property <flex-shrink> from [unset] to [2] at (1.5) should be [2.5]
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (-5) should be [0] assert_equals: expected "0 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <flex-shrink> from [1] to [2] at (1) should be [2]
FAIL CSS Transitions: property <flex-shrink> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (-5) should be [0] assert_equals: expected "0 " but got "- 4 "
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (-0.3) should be [0.7]
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (0) should be [1]
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (0.3) should be [1.3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (0.6) should be [1.6]
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (1) should be [2]
PASS CSS Transitions with transition: all: property <flex-shrink> from [1] to [2] at (1.5) should be [2.5]
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (-5) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (-5) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (0) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1.5 "
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (-5) should be [0] assert_equals: expected "0 " but got "1 "
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (-0.3) should be [0] assert_equals: expected "0 " but got "1 "
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (0) should be [0] assert_equals: expected "0 " but got "1 "
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (0.3) should be [0.3] assert_equals: expected "0.3 " but got "1 "
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (0.6) should be [0.6] assert_equals: expected "0.6 " but got "1 "
PASS CSS Transitions: property <flex-shrink> from [0] to [1] at (1) should be [1]
FAIL CSS Transitions: property <flex-shrink> from [0] to [1] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (-5) should be [0] assert_equals: expected "0 " but got "- 5 "
FAIL CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (-0.3) should be [0] assert_equals: expected "0 " but got "- 0.3 "
PASS CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (0) should be [0]
PASS CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (0.3) should be [0.3]
PASS CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (0.6) should be [0.6]
PASS CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (1) should be [1]
PASS CSS Transitions with transition: all: property <flex-shrink> from [0] to [1] at (1.5) should be [1.5]
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (-5) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (-0.3) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (0) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (0.3) should be [0.3] assert_equals: expected "0.3 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (0.6) should be [0.6] assert_equals: expected "0.6 " but got "1.5 "
FAIL CSS Animations: property <flex-shrink> from [0] to [1] at (1) should be [1] assert_equals: expected "1 " but got "1.5 "
PASS CSS Animations: property <flex-shrink> from [0] to [1] at (1.5) should be [1.5]
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (-5) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (-0.3) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (0) should be [0] assert_equals: expected "0 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (0.3) should be [0.3] assert_equals: expected "0.3 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (0.6) should be [0.6] assert_equals: expected "0.6 " but got "1.5 "
FAIL Web Animations: property <flex-shrink> from [0] to [1] at (1) should be [1] assert_equals: expected "1 " but got "1.5 "
PASS Web Animations: property <flex-shrink> from [0] to [1] at (1.5) should be [1.5]
